backend_django/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "backend_django"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
environs = "^9.5.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
six = "^1.16.0"
mailjet-rest = "^1.3.4"
celery = "^5.2.7"
drf-yasg = "^1.21.4"
djangorestframework-camel-case = "^1.3.0"
redis = "^4.4.0"
coverage = "^6.5.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
